requirements.txt
fastapi==0.112.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
uvicorn==0.30.5
SQLAlchemy==2.0.32
pymssql==2.3.1
passlib==1.7.4
bcrypt==4.0.1
msal==1.30.0
redis==5.0.8
